Remove logging dependecies in puppeteer-parse
