+++ packages/content-fetch/package.json
@@ -8,7 +8,8 @@
     "express": "^4.17.1",
     "@google-cloud/functions-framework": "^3.0.0",
     "@omnivore/puppeteer-parse": "^1.0.0",
-    "@sentry/serverless": "^6.13.3"
+    "@sentry/serverless": "^6.13.3",
+    "winston": "^3.3.3"
   },
   "devDependencies": {
     "chai": "^4.3.6",
