+++ packages/puppeteer-parse/package.json
@@ -4,19 +4,16 @@
   "description": "Accepts URL of the article and parses its content",
   "main": "index.js",
   "dependencies": {
-    "@google-cloud/logging-winston": "^5.1.1",
     "@google-cloud/storage": "^5.18.1",
     "@omnivore/content-handler": "1.0.0",
     "axios": "^0.27.2",
     "jsonwebtoken": "^8.5.1",
     "linkedom": "^0.14.9",
-    "luxon": "^2.3.1",
     "puppeteer-core": "^16.1.0",
     "puppeteer-extra": "^3.3.4",
     "puppeteer-extra-plugin-adblocker": "^2.13.5",
     "puppeteer-extra-plugin-stealth": "^2.11.1",
-    "underscore": "^1.13.4",
-    "winston": "^3.3.3"
+    "underscore": "^1.13.4"
   },
   "devDependencies": {
     "chai": "^4.3.6",
